+++ src/api_virtual.h
@@ -6,4 +6,4 @@
 int initVirtual();
 int getHeaterStatusVirtual();
 int setHeaterStatusVirtual(HeaterStatus power);
-int getTemperatureVirtual();
+float getTemperatureVirtual();
